tts-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
chatterbox-tts
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12
loguru>=0.7.0
soundfile>=0.12.1
librosa>=0.10.0
